This is android journal app, it uses sqlite to store user data, The user has to login via Google or firebase authentication to access the app functionalities. When you register with firebase all the user data is saved on firebase database. some features such as Forget password and new user registraion has been integrated in the app
This is android journal app, it uses sqlite to store user data, The user has to login via Google or firebase authentication to access the app functionalities. When you register with firebase all the user data is saved on firebase database. some features such as Forget password and new user registraion has been integrated in the app